run postinstall nspawn with loopdev if bootable
This enables nspawn blockdev access for the postinstall script when a
bootable image is being created.
(Needed in order to e.g. make changes to bootloader from within the
postinstall script.)

Also this commit deduplicates the code required for setting up these
nspawn parameters and moves it into a separate method, that is being
called from multiple places.