mkosi: add very simple image versioning concept
This adds a "image version" concept to mkos. Via the
ImageVersion=/--image-version= settings or the `mkosi.version` file a
version string may be defined. This is used for two things:

1. It is included in the (default) output filename. i.e. instead of
   "image.raw" the files will be called "image-0.1.raw" for a version of
   "0.1" and so on.

2. It is passed as IMAGE_VERSION to the build script (the script could
   use it and patch it into /etc/os-release or so for example)

(In case you wonder why this is called ImageVersion=/--image-version=
instead of just Version=/--version, that's because there's already a
command line switch for the latter that prints mkosi's own version
number)

Fixes: #626