mkosi: add support for building cpio archives
This may be useful to create initramfs images, see
https://github.com/keszybz/mkosi-initrd.

I also tried an implementation using libarchive. The code was a little
bit simpler, but unfortunately the generated archive wasn't to the
kernel's liking. I think both the compressio options *and* the
underlying cpio format were a bit different. I think we should just accept
that the kernel is only tested with the specific output produced by cpio
with some specific settings.