mkosi: call dracut/unified kernel image on the final image, not the cached image
We need to invoke dracut on the final image, not on the cached image.
Currently, it's done the other way round, in light of the fact that
dracut is the slowest part of the build. However that doesn't really
work, since the generated unified kernel image depends on the final
build.

The generated image after all will:

1. when verity is enabled receives the root hash which depends on the
   final fs image to be complete

2. when the image id/version logic is enabled the unified kernel image
   name will change with every build

3. we want to base the initrd on the resources installed in the OS
   during the build. Besides configuration this prominently includes a
   freshly build systemd. After all, one of the major usecases for mkosi
   was to be useful to build test images for systemd development. There
   we really want to ensure the systemd we just built is included in the
   initrd, and not an old version that happened to be around when the
   cache image was first created.

This reverts major parts of e4277c77618f47be6a5126a709f64ad48492807b