man: make files bold in list, adjust phrasing
We said that all files are optional *before* starting the list of files,
then for each individual file, and then again *after* the list. Let's
simplify this by describing files without mentioning that is is optional.
Similarly, we don't need to emphasize the fact that scripts are executable
for each script… readers generally know that.

Let's also make the terms bold, so it's easier to find the interesting
paragraph in this wall of text.

Also break into paragraphs more; remove underlining from "OS" — it's a
normal abbreviation and the markup was confusing; use "/" at the end
of directories; use more markup for paths; move the description of
mkosi.skeleton above mkosi.extra.

Co-authored-by: Jörg Behrmann <behrmann@physik.fu-berlin.de>