mkosi: include the ellipsis+dot in complete_step() in the format string
complete_step() already added an ellipsis and dot on its own (to the
opening message and to the cloing message, respectively), but in some
places we had "..." or "." in the format string anyway, so we eneded
up with six dots or four dots in output. But instead of fixing those
places, I think it's better to go in the opposite direction, and include
the terminator in the format string.

This is nicer because:
- It's just easier to see the complete format string including the
  terminator. The "saving" of a few characters by including it
  automatically isn't very important.
- If the line ends in a path, we should skip the terminator to make
  the path easier to select-and-paste. This is now trivial: just skip
  the terminator in the format string.
- If the step is "immediate" and we expect it to finish quickly and
  without any further output (like calling os.unlink() or such), we
  can skip the ellipsis.

While at it, let's change "..." to "…", it looks nicer and saves some
screen real estate.