man: refer to settings primarily by their config file name
This implements the first part of #769: the structure of the man page
is changed from:

Command Line Parameters
--foo
...
Command Line Parameters and their Settings File Counterparts
--foo Output Foo=
...

to:

Settings
[Output] section
Foo=
--foo
...

I think the new order is both easier to read _and_ easier to maintain.
E.g. the table with option name mappings was missing a few entries.

I dropped "(see above)" and "(see below)" in various places. The user
knows that they can search for a given option name, and we don't need
to tell them. And since the target of this "anchor" usually isn't right
below or right above, they will need to use search _anyway_.

An exception is when we wanted to refer to a description of a file
in the Files section. In those cases I changed "(see below)" to
"(see "Files" section below)" so the user actually knows what to look
for.

I inserted FIXMEs for various issues. Trying to fix everything at once
would lead to an unreviable commit…

The regeneration of man/ is not included in the commits, because it's
annoying to constantly fix this up when rebasing patches. I'll do it
one go at the end.