mkosi: use Path for path-like objects
The Path object generally works better for the kind of operations that
we need to do. In some cases it can be a bit awkward — for example just
appending a suffix requires some annoying gyrations. But overall, the
amount of code text is reduced and I think it's easier to read.

Some bugs were fixed on the way: in particular dedent() was missed in
when writing some configuration files (90-mkosi-vmlinuz-add.hook,
60-mkosi-vmlinuz-remove.hook).

The intent is to have no functional change, though with this amount of
churn some regression is likely… I tested this by building a bunch
of different projects with diverse configurations, and at least mkosi
doesn't crash, though I didn't check the produced images in detail.

An additional advantage is that we have stronger typing: before many
functions took a bunch of 'str' arguments, and now the paths at least
have a distinct type.