mkosi: delete overlayfs whiteout files
Overlay uses char(0,0) files to mark items that are present in the lower
layers, but were removed in the upper layer. At least for the case
of sysexts, we want to get rid of those. Maybe if other uses of this
feature appear, we might want to make this removal optional.

Hat tip to @brau_ner for explaining what those files are.