Use unittest instead of pytest for new tests
Hacking on mkosi on older distributions becomes a lot easier when you
don't have any external library dependencies. You can simply get an
updated version of Python without the need to maintain external libraries
for that Python version.

This benefit seems substantial enough to me to justify using unittest
over pytest as it's part of the standard library. Arguably the biggest
loss is that we can't use "assert" anymore but have to use the assert
methods of unittest but this shouldn't be a huge deal.

I didn't touch the config tests since the plan is to remove them once
we add some simpler argparsing tests in the near future.

We can still use pytest to run all the tests as it supports running
unittest based tests as well.